written functions for send data messages from server
