+++ js/validation.js
@@ -1,21 +1,13 @@
-import {showAlert} from './util.js';
-import {sendData} from './api.js';
-
 const MAX_HASHTAG = 5;
 const MAX_LENGTH_DESCRIPTION = 140;
 const ERROR_DESCRIPTION = 'не более 140 символов';
 const ERROR_HASHTAG = 'Хештег содержит в начале #, состоит только из букв, чисел, без пробелов, спецсимволов, символов пунктуации, эмодзи и т.д.';
 
 const hashtagRegex = /^#[a-zа-яё\d]{1,19}$/i;
-const SubmitButtonText = {
-  IDLE: 'Опубликовать',
-  SENDING: 'Опубликовываю...'
-};
 
 const pictureFormElement = document.querySelector('.img-upload__form');
 const textHashtagElement = pictureFormElement.querySelector('.text__hashtags');
 const textDescriptionElement = pictureFormElement.querySelector('.text__description');
-const submitButtonElement = document.querySelector('.img-upload__submit');
 
 const pristine = new Pristine(pictureFormElement, {
   classTo: 'img-upload__field-wrapper',
@@ -42,30 +34,4 @@ pristine.addValidator(
   ERROR_HASHTAG
 );
 
-const blockSubmitButton = () => {
-  submitButtonElement.disabled = true;
-  submitButtonElement.textContent = SubmitButtonText.SENDING;
-};
-
-const unblockSubmitButton = () => {
-  submitButtonElement.disabled = false;
-  submitButtonElement.textContent = SubmitButtonText.IDLE;
-};
-
-const onFormSubmit = (onSuccess) => {
-  pictureFormElement.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-    const isValid = pristine.validate();
-    if (isValid) {
-      blockSubmitButton();
-      sendData(new FormData(evt.target))
-        .then(onSuccess)
-        .catch((err) => {
-          showAlert(err.message);
-        })
-        .finally(unblockSubmitButton);
-    }
-  });
-};
-
-export {onFormSubmit};
+export {pristine};
